@@ -7,13 +7,15 @@
     "serve": "vite preview"
   },
   "dependencies": {
+    "@vitejs/plugin-vue": "^2.2.2",
     "autoprefixer": "^10.4.2",
+    "eslint": "^8.9.0",
+    "eslint-plugin-vue": "^8.4.1",
     "postcss": "^8.4.6",
     "sass": "^1.49.7",
     "tailwindcss": "^3.0.18",
-    "vite-plugin-vue2": "^1.4.0",
-    "vue": "^2.6.12",
-    "vue-router": "^3.5.3"
+    "vue": "^3.2.31",
+    "vue-router": "^4.0.12"
   },
   "devDependencies": {
     "vite": "^2.0.5"
