@@ -1,7 +1,9 @@
-const { createVuePlugin } = require("vite-plugin-vue2");
 const path = require("path");
+import { defineConfig } from "vite";
+import vue from "@vitejs/plugin-vue";
 
-module.exports = {
+// https://vitejs.dev/config/
+export default defineConfig({
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
@@ -12,5 +14,5 @@ module.exports = {
       images: path.resolve(__dirname, "./src/assets/images"),
     },
   },
-  plugins: [createVuePlugin()],
-};
+  plugins: [vue()],
+});
